requirements.txt
annotated-types==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.3.3 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.28.2 ; python_version >= "3.12" and python_version < "4.0"
idna==3.6 ; python_version >= "3.12" and python_version < "4.0"
kubernetes==29.0.0 ; python_version >= "3.12" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.16.3 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.6.4 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.17.2 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
pyyaml[libyaml]==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
requests-oauthlib==1.4.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.12" and python_version < "4.0"
rich==13.7.1 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
shellingham==1.5.4 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
typer[all]==0.9.0 ; python_version >= "3.12" and python_version < "4.0"
types-pyyaml==6.0.12.20240311 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.10.0 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
websocket-client==1.7.0 ; python_version >= "3.12" and python_version < "4.0"
